requirements.txt
ImageHash==4.3.1
numpy==1.25.2
Pillow==10.0.0
PyWavelets==1.4.1
scipy==1.11.1
